Add shared spherical harmonic helpers in spherical utils
